refactor: seperate friendships to own app
