+++ backend/movement/urls.py
@@ -1,6 +1,5 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from rest_framework.routers import DefaultRouter
 from .views import MuscleViewSet, MovementViewSet
 
 '''
